hotfix: 기본 프로필 이미지 삭제되지 않도록 수정

+++ src/main/java/com/acon/server/global/external/s3/S3Adapter.java
@@ -109,7 +109,9 @@ public class S3Adapter {
     }
 
     public void deleteFile(String fileUrl) {
-        amazonS3.deleteObject(bucketName, getFileKey(fileUrl));
+        if (!fileUrl.equals(basicProfileImageUrl)) {
+            amazonS3.deleteObject(bucketName, getFileKey(fileUrl));
+        }
     }
 
     private String getFileKey(String fileUrl) {
